make a cachemixin
